dom/base/test/test_bug557892.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=557892
-->
<head>
  <title>Test for Bug 557892</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=557892">Mozilla Bug 557892</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 557892 **/

SimpleTest.waitForExplicitFinish();

window.open("file_bug557892.html", "", "width=100,height=100");

function done() {
  ok(true, "Should not crash");
  setTimeout(SimpleTest.finish, 0);
}

</script>
</pre>
</body>
</html>
